Add a vector-based application icon
This new icon was drawn using Inkscape and is available in SVG format.
It has a slightly different appearance from the previous icon.

The icon is also available in PNG from 32×32 to 256×256, so it can be
installed system-wide on Linux.

The Windows icon file now includes sizes up to 256×256 for a better
appearance on hiDPI displays.

This closes #39.